Revert "Ramaanibal"
