--- src/components/Developer.tsx
@@ -1,11 +1,0 @@
-class Developer{
-    id: number;
-    nombre: string;
-    constructor(
-        id = 0,
-        nombre = "vacio"
-        )
-    {
-    this.id = id,
-    this.nombre = nombre}    
-}
