+++ src/routes/App.css
@@ -1,23 +1,19 @@
-#custom-search-bar {
-  margin-block-end: 10px;
-  width: 80%; /* Adjust the width as needed */
-  height: 60px;
-  box-sizing: border-box;
-  border: 4px solid #ccc;
-  border-radius: 30px;
-  background-color: #f8f8f8;
-  margin: 10 auto;
-}
-
+#search-bar{
+    margin-block-end: 10px;
+    width: 100%;
+    height: 60px;
+    box-sizing: border-box;
+    border: 4px solid #ccc;
+    border-radius: 30px;
+    background-color: #f8f8f8;
+    margin: 10 auto;
+    
+  }
   .space-taker{
       height: 20px;
   }
   
-body{
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  background: #17223b;
-  color: #f0f0f0;
-}
+  
 #body-listado{
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
     background: #17223b;
@@ -103,35 +99,3 @@ body{
     color: #6610f2;
   }
   
-  .reportes-por-producto{
-
-    width: 100%;
- 
-    margin-top: 70px;
-
-  }
- 
- 
-.bugs-list-container {
-  display: flex;
-  flex-direction: column;
-}
-.dropdown-container{
-  margin-left:2.5%;
-}
-.search-container{
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  background: #17223b;
-  color: #f0f0f0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: 80vh;     
-}
-.table-search-container{
-  height: 170%; /* Adjust the height as needed */
-  width: 129%; /* Optional: Adjust the width as needed */
-  margin-left: -16.5%;
-
-}
