+++ src/routes/Ingresa.css
@@ -12,10 +12,9 @@
   .space-taker{
       height: 25px;
   }
-  #arriba-de-form{
-
-  }
-  #form-bug {
+  
+  
+  body {
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
     background: #17223b;
     color: #f0f0f0;
@@ -24,7 +23,6 @@
     justify-content: center;
     min-height: 90vh;
   }
-
   
   #form-react {
     padding: 30px 20px;
